requirements/base.txt
Django==5.2.4
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.1
social-auth-app-django==5.5.1
drf-spectacular
django-environ
Pillow==10.4.0          # Image processing
pymysql
celery[redis]            # optional async tasks
google-api-python-client==2.177.0
google-auth==2.28.1
django-cors-headers==4.7.0
django-storages
google-cloud-storage
